+++ Fin.lua
@@ -1,14 +1,11 @@
 -- Fin.lua
 
-Fin = Gamestate.new(); 
-function Fin:init()
-	self.background = love.graphics.newImage("assets/FIN.png");
-end
+Fin = Gamestate.new()
+local finBg = love.graphics.newImage("assets/FIN.png")
+
 function Fin:draw()
-	love.graphics.draw(Fin.background, 0, 0);	
+	love.graphics.draw(finBg, 0, 0)
 end
 function Fin:keypressed(k)
-	if k == 'q' then 
-		love.event.push('quit');
-	end
+	if k == 'q' then love.event.push('quit') end
 end